Moved test files
1. Moved `tests/1/` and `tests/2/` to `src/1/__tests__/` and `src/2/__tests__/` respectively
2. Updated relative imports
3. Updated test file paths
