+++ src/2/__tests__/json_checker/fail6.json
