+++ src/2/__tests__/json_checker/pass3.json
